ml_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
tensorflow-cpu==2.15.0
pillow==10.2.0
numpy==1.26.3
